refactor: move slash form mounting into command layer
